refactor(ui): refactor requirements table to support hot module replacement
