+++ ui/src/pages/Calculator/components/Requirements/RequirementRow.tsx
@@ -9,35 +9,12 @@ import {
     NumberColumnCell,
     TextColumnCell,
 } from "./styles";
-
-export type SortableFields = {
-    amount: number;
-    workers: number;
-};
-
-export type SingleCreatorRequirementsTableRow = SortableFields & {
-    name: string;
-    creator: string;
-};
-
-export type MultipleCreatorRequirementsTableRow = Omit<
+import {
+    MultipleCreatorRequirementsTableRow,
+    RequirementsTableRow,
     SingleCreatorRequirementsTableRow,
-    "creator"
-> & {
-    isExpanded: boolean;
-    creatorBreakdownRows: Omit<SingleCreatorRequirementsTableRow, "name">[];
-};
-
-export type RequirementsTableRow =
-    | SingleCreatorRequirementsTableRow
-    | MultipleCreatorRequirementsTableRow;
-
-function isSingleCreatorRow(
-    row: RequirementsTableRow
-): row is SingleCreatorRequirementsTableRow {
-    const casted = row as SingleCreatorRequirementsTableRow;
-    return casted.creator !== undefined;
-}
+    isSingleCreatorRow,
+} from "./types";
 
 type RequirementsRowProps = {
     row: RequirementsTableRow;
@@ -129,4 +106,4 @@ function RequirementRow({ row, toggleCreatorBreakdown }: RequirementsRowProps) {
     );
 }
 
-export { RequirementRow, isSingleCreatorRow };
+export { RequirementRow };
